+++ src/components/menu/buttons/SortBtn.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { SortType, SortedAt } from '../../../app/types/folder';
 import { useAppDispatch, useAppSelector } from '../../../app/hooks';
 import { changeSortTypeOfFolder } from '../../../app/actions/folder';
-import useClickOutOfElement from '../../../hooks/useElementClickOutOfArea';
+import useClickOutOfElement from '../../../hooks/useClickOutOfElement';
 import useDelay from '../../../hooks/useDelay';
 import usePathname from '../../../hooks/usePathname';
